# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ModaBrasileira",
        "summary": "Routes of app",
        "description": "API under development",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Jhonattan Rocha da Silva",
            "url": "http://www.example.com/contact/",
            "email": "jhonattab246rocha@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "identifier": "MIT"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/crud/subcategories/": {
            "post": {
                "summary": "Create Subcategory",
                "operationId": "create_subcategory_crud_subcategories__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubCategoryBase"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubCategoryCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Read Subcategories",
                "operationId": "read_subcategories_crud_subcategories__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "filters",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Filters"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubCategory"
                                    },
                                    "title": "Response Read Subcategories Crud Subcategories  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/crud/subcategories/{subcategory_id}": {
            "get": {
                "summary": "Read Subcategory",
                "operationId": "read_subcategory_crud_subcategories__subcategory_id__get",
                "parameters": [
                    {
                        "name": "subcategory_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Subcategory Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubCategory"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Subcategory",
                "operationId": "update_subcategory_crud_subcategories__subcategory_id__put",
                "parameters": [
                    {
                        "name": "subcategory_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Subcategory Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubCategoryCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubCategoryCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Subcategory",
                "operationId": "delete_subcategory_crud_subcategories__subcategory_id__delete",
                "parameters": [
                    {
                        "name": "subcategory_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Subcategory Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/crud/people": {
            "post": {
                "summary": "Create People",
                "operationId": "create_people_crud_people_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonBase"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Read People",
                "operationId": "read_people_crud_people_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Person"
                                    },
                                    "title": "Response Read People Crud People Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/crud/people/{person_id}": {
            "get": {
                "summary": "Read People",
                "operationId": "read_people_crud_people__person_id__get",
                "parameters": [
                    {
                        "name": "person_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Person Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update People",
                "operationId": "update_people_crud_people__person_id__put",
                "parameters": [
                    {
                        "name": "person_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Person Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete People",
                "operationId": "delete_people_crud_people__person_id__delete",
                "parameters": [
                    {
                        "name": "person_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Person Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Person": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cpf": {
                        "type": "string",
                        "title": "Cpf"
                    },
                    "tel": {
                        "type": "string",
                        "title": "Tel"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "cpf",
                    "tel",
                    "email",
                    "address",
                    "id"
                ],
                "title": "Person"
            },
            "PersonBase": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cpf": {
                        "type": "string",
                        "title": "Cpf"
                    },
                    "tel": {
                        "type": "string",
                        "title": "Tel"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "cpf",
                    "tel",
                    "email",
                    "address"
                ],
                "title": "PersonBase"
            },
            "PersonCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cpf": {
                        "type": "string",
                        "title": "Cpf"
                    },
                    "tel": {
                        "type": "string",
                        "title": "Tel"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "cpf",
                    "tel",
                    "email",
                    "address",
                    "id"
                ],
                "title": "PersonCreate"
            },
...
            "SubCategory": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id_cat": {
                        "type": "integer",
                        "title": "Id Cat"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id_cat",
                    "id"
                ],
                "title": "SubCategory"
            },
            "SubCategoryBase": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id_cat": {
                        "type": "integer",
                        "title": "Id Cat"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id_cat"
                ],
                "title": "SubCategoryBase"
            },
            "SubCategoryCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id_cat": {
                        "type": "integer",
                        "title": "Id Cat"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id_cat",
                    "id"
                ],
                "title": "SubCategoryCreate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/crud/token"
                    }
                }
            }
        }
    }
}